// src/interfaces/SpotifyTrackFeatures.ts
interface SpotifyTrackFeatures {
	id: string;
	danceability: number;
	energy: number;
	key: number;
	loudness: number;
	speechiness: number;
	acousticness: number;
	instrumentalness: number;
	liveness: number;
	valence: number;
	tempo: number;
	duration: number;
	timeSignature: number;
}

export default SpotifyTrackFeatures;